fix(web_view): export viewport from a server layout (Chromium ignores DOM mutation)
The previous attempt set the viewport meta from a useEffect, but
Android Chromium evaluates `<meta name=viewport>` exactly once at
initial layout — any later DOM mutation is ignored, so on a real
device the WebView kept Next.js's default
`width=device-width, initial-scale=1` and the four downstream issues
(pinch zoom, sticky headers off-screen, body overflow, keyboard slide-
in) all survived.

Split the web_view layout into:
- layout.tsx (server component) exporting `viewport: Viewport` so
  Next.js renders the strict native-app meta into the served HTML.
- WebViewClientLayout.tsx for every existing hook / effect / bridge
  wiring (unchanged behaviour).

`viewport-fit: cover` keeps `env(safe-area-inset-*)` resolving to
real notch values on iOS; `userScalable: false` and `maximumScale: 1`
disable pinch zoom on both platforms; `width: device-width` forces
layout viewport == visual viewport on Android Chromium.
